week1/instant/requirements.txt
uvicorn
fastapi
google-genai